+++ rocketcea/examples/chk_cea_long_name.py
@@ -1,9 +1,9 @@
-from rocketcea.cea_obj import CEA_Obj, get_rocketcea_data_dir, set_rocketcea_data_dir
+from rocketcea.cea_obj import CEA_Obj, get_rocketcea_data_dir #, set_rocketcea_data_dir
 
 test_dir = r'D:\temp\cea w spaces'
 print( 'ROCKETCEA_DATA_DIR =', get_rocketcea_data_dir() )
 
-set_rocketcea_data_dir( test_dir )
+# set_rocketcea_data_dir( test_dir )
 
 print( 'ROCKETCEA_DATA_DIR =', get_rocketcea_data_dir() )